+++ integration-tests/deployment/ccip/add_chain.go
@@ -15,8 +15,7 @@ import (
 	"github.com/smartcontractkit/chainlink-common/pkg/types/ccipocr3"
 
 	"github.com/smartcontractkit/chainlink/integration-tests/deployment"
-	cctypes "github.com/smartcontractkit/chainlink/v2/core/capabilities/ccip/types"
-
+	"github.com/smartcontractkit/chainlink/v2/core/capabilities/ccip/types"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/ccip_home"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/fee_quoter"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/onramp"
@@ -29,7 +28,10 @@ func NewChainInboundProposal(
 	state CCIPOnChainState,
 	homeChainSel uint64,
 	newChainSel uint64,
+	feedChainSel uint64,
+	tokenConfig TokenConfig,
 	sources []uint64,
+	rmnHomeAddress common.Address,
 ) (*timelock.MCMSWithTimelockProposal, error) {
 	// Generate proposal which enables new destination (from test router) on all source chains.
 	var batches []timelock.BatchChainOperation
@@ -117,6 +119,35 @@ func NewChainInboundProposal(
 	if err != nil {
 		return nil, err
 	}
+	newDONArgs, err := BuildOCR3ConfigForCCIPHome(
+		e.Logger,
+		state.Chains[newChainSel].OffRamp,
+		e.Chains[newChainSel],
+		feedChainSel,
+		tokenConfig.GetTokenInfo(e.Logger, state.Chains[newChainSel].LinkToken),
+		nodes.NonBootstraps(),
+		rmnHomeAddress,
+	)
+	if err != nil {
+		return nil, err
+	}
+	latestDon, err := LatestCCIPDON(state.Chains[homeChainSel].CapabilityRegistry)
+	if err != nil {
+		return nil, err
+	}
+	commitConfig, ok := newDONArgs[types.PluginTypeCCIPCommit]
+	if !ok {
+		return nil, fmt.Errorf("missing commit plugin in ocr3Configs")
+	}
+	donID := latestDon.Id + 1
+	addDonOp, err := SetCandidateCommitPluginWithAddDonOps(
+		donID, commitConfig,
+		state.Chains[homeChainSel].CapabilityRegistry,
+		nodes.NonBootstraps(),
+	)
+	if err != nil {
+		return nil, err
+	}
 	timelockAddresses, metaDataPerChain, err := BuildProposalMetadata(state, append(chains, homeChainSel))
 	if err != nil {
 		return nil, err
@@ -130,6 +161,7 @@ func NewChainInboundProposal(
 				Data:  addChain.Data(),
 				Value: big.NewInt(0),
 			},
+			addDonOp,
 		},
 	})
 	return timelock.NewMCMSWithTimelockProposal(
@@ -172,7 +204,7 @@ func AddDonAndSetCandidateForCommitProposal(
 	if err != nil {
 		return nil, err
 	}
-	commitConfig, ok := newDONArgs[cctypes.PluginTypeCCIPCommit]
+	commitConfig, ok := newDONArgs[types.PluginTypeCCIPCommit]
 	if !ok {
 		return nil, fmt.Errorf("missing commit plugin in ocr3Configs")
 	}
@@ -200,103 +232,6 @@ func AddDonAndSetCandidateForCommitProposal(
 		[]timelock.BatchChainOperation{{
 			ChainIdentifier: mcms.ChainIdentifier(homeChainSel),
 			Batch:           []mcms.Operation{addDonOp},
-		}},
-		timelock.Schedule,
-		"0s", // TODO: Should be parameterized.
-	)
-}
-
-// SetCandidateExecPluginProposal calls setCandidate on the CCIPHome for setting up OCR3 exec Plugin config for the new chain.
-func SetCandidateExecPluginProposal(
-	state CCIPOnChainState,
-	e deployment.Environment,
-	nodes deployment.Nodes,
-	homeChainSel, feedChainSel, newChainSel uint64,
-	tokenConfig TokenConfig,
-	rmnHomeAddress common.Address,
-) (*timelock.MCMSWithTimelockProposal, error) {
-	newDONArgs, err := BuildOCR3ConfigForCCIPHome(
-		e.Logger,
-		state.Chains[newChainSel].OffRamp,
-		e.Chains[newChainSel],
-		feedChainSel,
-		tokenConfig.GetTokenInfo(e.Logger, state.Chains[newChainSel].LinkToken),
-		nodes.NonBootstraps(),
-		rmnHomeAddress,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	execConfig, ok := newDONArgs[cctypes.PluginTypeCCIPExec]
-	if !ok {
-		return nil, fmt.Errorf("missing exec plugin in ocr3Configs")
-	}
-
-	setCandidateMCMSOps, err := SetCandidateExecPluginOps(
-		execConfig,
-		state.Chains[homeChainSel].CapabilityRegistry,
-		state.Chains[homeChainSel].CCIPHome,
-		newChainSel,
-		nodes.NonBootstraps(),
-	)
-
-	if err != nil {
-		return nil, err
-	}
-	timelockAddresses, metaDataPerChain, err := BuildProposalMetadata(state, []uint64{homeChainSel})
-	if err != nil {
-		return nil, err
-	}
-	return timelock.NewMCMSWithTimelockProposal(
-		"1",
-		2004259681, // TODO: should be parameterized and based on current block timestamp.
-		[]mcms.Signature{},
-		false,
-		metaDataPerChain,
-		timelockAddresses,
-		"SetCandidate for execution",
-		[]timelock.BatchChainOperation{{
-			ChainIdentifier: mcms.ChainIdentifier(homeChainSel),
-			Batch:           setCandidateMCMSOps,
-		}},
-		timelock.Schedule,
-		"0s", // TODO: Should be parameterized.
-	)
-}
-
-// PromoteCandidateProposal generates a proposal to call promoteCandidate on the CCIPHome through CapReg.
-// This needs to be called after SetCandidateProposal is executed.
-func PromoteCandidateProposal(
-	state CCIPOnChainState,
-	homeChainSel, newChainSel uint64,
-	nodes deployment.Nodes,
-) (*timelock.MCMSWithTimelockProposal, error) {
-	promoteCandidateOps, err := PromoteCandidateOps(
-		state.Chains[homeChainSel].CapabilityRegistry,
-		state.Chains[homeChainSel].CCIPHome,
-		newChainSel,
-		nodes.NonBootstraps(),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	timelockAddresses, metaDataPerChain, err := BuildProposalMetadata(state, []uint64{homeChainSel})
-	if err != nil {
-		return nil, err
-	}
-	return timelock.NewMCMSWithTimelockProposal(
-		"1",
-		2004259681, // TODO: should be parameterized and based on current block timestamp.
-		[]mcms.Signature{},
-		false,
-		metaDataPerChain,
-		timelockAddresses,
-		"promoteCandidate for commit and execution",
-		[]timelock.BatchChainOperation{{
-			ChainIdentifier: mcms.ChainIdentifier(homeChainSel),
-			Batch:           promoteCandidateOps,
 		}},
 		timelock.Schedule,
 		"0s", // TODO: Should be parameterized.
